+++ tests/test_textdirectory.py
@@ -8,7 +8,8 @@ from click.testing import CliRunner
 
 from textdirectory.textdirectory import TextDirectory
 from textdirectory.transformations import transformation_remove_non_ascii, transformation_remove_non_alphanumerical, \
-    transformation_to_leetspeak, transformation_crude_spellchecker, transformation_remove_weird_tokens
+    transformation_to_leetspeak, transformation_crude_spellchecker, transformation_remove_stopwords, \
+    transformation_remove_htmltags, transformation_remove_weird_tokens
 from textdirectory import cli
 
 
@@ -33,14 +34,14 @@ def test_iterator():
 
 def test_simpple_aggregations():
     """Test the simplest form of aggregation."""
-    td = TextDirectory(directory='data/testdata/')
+    td = TextDirectory(directory='textdirectory/data/testdata/')
     td.load_files(True, 'txt')
     assert len(td.aggregate_to_memory()) == 4083
 
 
 def test_filter_by_chars_outliers():
     """Test the outlier filter."""
-    td = TextDirectory(directory='data/testdata/')
+    td = TextDirectory(directory='textdirectory/data/testdata/')
     td.load_files(True, 'txt')
     td.filter_by_chars_outliers(1)
     assert len(td.aggregation) == 7
@@ -48,15 +49,15 @@ def test_filter_by_chars_outliers():
 
 def test_filter_by_similar_documents():
     """Test the similarity filter."""
-    td = TextDirectory(directory='data/testdata/')
+    td = TextDirectory(directory='textdirectory/data/testdata/')
     td.load_files(True, 'txt')
-    td.filter_by_similar_documents(reference_file='data/testdata/level_2/Text_E.txt', threshold=0.7)
+    td.filter_by_similar_documents(reference_file='textdirectory/data/testdata/level_2/Text_E.txt', threshold=0.7)
     assert len(td.aggregation) == 2
 
 
 def test_filter_by_max_filesize():
     """Test the filesize (max) filter."""
-    td = TextDirectory(directory='data/testdata/')
+    td = TextDirectory(directory='textdirectory/data/testdata/')
     td.load_files(True, 'txt')
     td.filter_by_max_filesize(max_kb=1)
     assert len(td.aggregation) == 7
@@ -64,7 +65,7 @@ def test_filter_by_max_filesize():
 
 def test_filter_by_min_filesize():
     """Test the filesize (min) filter."""
-    td = TextDirectory(directory='data/testdata/')
+    td = TextDirectory(directory='textdirectory/data/testdata/')
     td.load_files(True, 'txt')
     td.filter_by_min_filesize(min_kb=2)
     assert len(td.aggregation) == 1
@@ -72,21 +73,27 @@ def test_filter_by_min_filesize():
 
 def test_transformation_remove_nl():
     """Test the remove_nl transformation."""
-    td = TextDirectory(directory='data/testdata/')
+    td = TextDirectory(directory='textdirectory/data/testdata/')
     td.load_files(True, 'txt')
     td.stage_transformation(['transformation_remove_nl'])
     assert '\n' not in td.aggregate_to_memory()
 
 
+def test_transformation_remove_htmltags():
+    """Test the remove htmltags transformation."""
+    test_string = '<html><body>This <span id="1">is</span> a <em>test</em></body></html>'
+    assert transformation_remove_htmltags(test_string) == 'This is a test'
+
+
 def test_transformation_uppercase():
     """Test the uppercase transformation."""
-    td = TextDirectory(directory='data/testdata/')
+    td = TextDirectory(directory='textdirectory/data/testdata/')
     td.load_files(True, 'txt')
     td.stage_transformation(['transformation_uppercase'])
     assert td.aggregate_to_memory().isupper()
 
 
-def test_transformation_remove_non_ascii():
+def test_transformation_remove_non_ascii_hard():
     """Test the remove non-ascii transformation."""
     test_string = 'This is a @ test string ~ containing non-ascii characters such as 😁.'
     assert transformation_remove_non_ascii(test_string) == 'This is a @ test string ~ containing non-ascii characters such as .'
@@ -116,17 +123,33 @@ def test_transformation_remove_weird_tokens():
     assert transformation_remove_weird_tokens(test_string, remove_double_space=True) == 'Hello World!'
 
 
-#def test_transformation_postag():
-#    """Test the postag transformation."""
-#    td = TextDirectory(directory='data/testdata/')
-#    td.load_files(True, 'txt')
-#    td.stage_transformation(['transformation_postag'])
-#    assert 'NN' in td.aggregate_to_memory()
+def test_transformation_remove_stopwords():
+    """Test the remove stopwords transformation."""
+    test_string = 'There is a house on the hill.'
+    assert transformation_remove_stopwords(test_string) == 'There is house hill.'
+
+
+def test_transformation_test_arguments():
+    """Test whether we can pass arguments to transformations."""
+    td = TextDirectory(directory='textdirectory/data/testdata/')
+    td.load_files(True, 'txt')
+    td.stage_transformation(['transformation_remove_stopwords', 'internal', 'en', 'en_core_web_sm',
+                             'dolor,dolore,dolores'])
+    text = td.aggregate_to_memory()
+    assert 'dolor' not in text
+
+
+def test_transformation_postag():
+    """Test the postag transformation."""
+    td = TextDirectory(directory='textdirectory/data/testdata/')
+    td.load_files(True, 'txt')
+    td.stage_transformation(['transformation_postag'])
+    assert 'NN' in td.aggregate_to_memory()
 
 
 def test_tabulation(capsys):
     """Test the tabulation."""
-    td = TextDirectory(directory='data/testdata/')
+    td = TextDirectory(directory='textdirectory/data/testdata/')
     td.load_files(True, 'txt')
     td.print_aggregation()
     out, err = capsys.readouterr()
@@ -145,7 +168,7 @@ def test_print_pipeline(capsys):
 
 def test_transform_to_memory():
     """Test the in memory transformation."""
-    td = TextDirectory(directory='data/testdata/')
+    td = TextDirectory(directory='textdirectory/data/testdata/')
     td.load_files(True, 'txt')
     td.transform_to_memory()
     assert len(list(td.get_aggregation())[0]['transformed_text']) > 0