+++ textdirectory/data/stopwords/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+
+"""Top-level package for textdirectory."""
+
+__author__ = """Ingo Kleiber"""
+__email__ = 'ingo@kleiber.me'
+__version__ = '0.2.0'
+
+from textdirectory import *
+from textdirectory.textdirectory import *